// src/conversation/conversation-state.enum.ts
export enum ConversationState {
  IDLE = 'IDLE',
  WAITING_AMOUNT = 'WAITING_AMOUNT',
  WAITING_PROVIDER = 'WAITING_PROVIDER',
  WAITING_CATEGORY = 'WAITING_CATEGORY',
  WAITING_REASON = 'WAITING_REASON',
  WAITING_METHOD = 'WAITING_METHOD',
  WAITING_RECEIPT = 'WAITING_RECEIPT',
  WAITING_VOICE_EXPENSE = 'WAITING_VOICE_EXPENSE',
  WAITING_CONFIRMATION = 'WAITING_CONFIRMATION',
  EDITING_FIELD = 'EDITING_FIELD',
  WAITING_QUESTION = 'WAITING_QUESTION',
}